+++ app/Http/Controllers/MasterKelasController.php
@@ -9,6 +9,7 @@ use App\Dosen;
 use App\User;
 use App\Jurusan;
 use DB;
+use Illuminate\Support\Facades\Auth;
 
 class MasterKelasController extends Controller
 {
@@ -21,13 +22,22 @@ class MasterKelasController extends Controller
 
     public function index()
     {
-        $kelas = DB::table('kelas')
-        ->join('matkuls', 'kelas.matkul_id', '=', 'matkuls.id')
-        ->join('dosens', 'kelas.dosen_id', '=', 'dosens.id')
-        ->select('kelas.*', 'matkuls.nama as nama_matkul', 'dosens.nama as nama_dosen')
-        ->get();
+        $user = Auth::user();
+        if($user['status'] == 'admin')
+        {
+            $kelas = DB::table('kelas')
+            ->join('matkuls', 'kelas.matkul_id', '=', 'matkuls.id')
+            ->join('dosens', 'kelas.dosen_id', '=', 'dosens.id')
+            ->select('kelas.*', 'matkuls.nama as nama_matkul', 'dosens.nama as nama_dosen')
+            ->get();
 
-        return view('admin.master_kelas', compact('kelas'));
+            return view('admin.master_kelas', compact('kelas'));
+        }
+        else
+        {
+            return view('noaccess');
+        }
+        
     }
 
     /**
@@ -37,10 +47,19 @@ class MasterKelasController extends Controller
      */
     public function create()
     {
-        $jurusans = Jurusan::all();
-        $dosens = Dosen::all();
-        $matkuls = Matkul::all();
-        return view('admin.input_kelas', compact('jurusans','dosens', 'matkuls'));
+        $user = Auth::user();
+        if($user['status'] == 'admin')
+        {
+            $jurusans = Jurusan::all();
+            $dosens = Dosen::all();
+            $matkuls = Matkul::all();
+            return view('admin.input_kelas', compact('jurusans','dosens', 'matkuls'));
+        }
+        else
+        {
+            return view('noaccess');
+        }
+        
     }
 
     /**
@@ -51,20 +70,29 @@ class MasterKelasController extends Controller
      */
     public function store(Request $request)
     {
-        $kelas = new Kelas([ 
-          'hari' => $request->get('hari'),
-          'jam_masuk' => $request->get('jammasuk'),
-          'jam_keluar' => $request->get('jamkeluar'),
-          'kp' => $request->get('kp'),
-          'kapasitas' => $request->get('kapasitas'),
-          'ruang' => $request->get('ruang'),
-          'matkul_id' => $request->get('matkul'),
-          'dosen_id' => $request->get('dosen'),
-      ]);
+        $user = Auth::user();
+        if($user['status'] == 'admin')
+        {
+            $kelas = new Kelas([ 
+              'hari' => $request->get('hari'),
+              'jam_masuk' => $request->get('jammasuk'),
+              'jam_keluar' => $request->get('jamkeluar'),
+              'kp' => $request->get('kp'),
+              'kapasitas' => $request->get('kapasitas'),
+              'ruang' => $request->get('ruang'),
+              'matkul_id' => $request->get('matkul'),
+              'dosen_id' => $request->get('dosen'),
+          ]);
 
-        $kelas->save();
+            $kelas->save();
 
-        return redirect('master_kelas');
+            return redirect('master_kelas');
+        }
+        else
+        {
+            return view('noaccess');
+        }
+        
     }
 
     /**
@@ -86,14 +114,23 @@ class MasterKelasController extends Controller
      */
     public function edit($id)
     {
-        $kelas = Kelas::find($id);
-        $matkuls = Matkul::find($kelas->matkul_id);
-        $dosens = Dosen::find($kelas->dosen_id);
+        $user = Auth::user();
+        if($user['status'] == 'admin')
+        {
+            $kelas = Kelas::find($id);
+            $matkuls = Matkul::find($kelas->matkul_id);
+            $dosens = Dosen::find($kelas->dosen_id);
 
-        $fullmatkuls = Matkul::all();
-        $fulldosens = Dosen::all();
+            $fullmatkuls = Matkul::all();
+            $fulldosens = Dosen::all();
+            
+            return view('admin.edit_kelas', compact('fullmatkuls','fulldosens','dosens','matkuls','kelas'));
+        }
+        else
+        {
+            return view('noaccess');
+        }
         
-        return view('admin.edit_kelas', compact('fullmatkuls','fulldosens','dosens','matkuls','kelas'));
     }
 
     /**
@@ -105,20 +142,29 @@ class MasterKelasController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $kelas = Kelas::find($id);
-        $kelas->hari = $request->get('hari');
-        $kelas->hari = $request->get('hari');
-        $kelas->jam_masuk = $request->get('jammasuk');
-        $kelas->jam_keluar = $request->get('jamkeluar');
-        $kelas->kp = $request->get('kp');
-        $kelas->kapasitas = $request->get('kapasitas');
-        $kelas->ruang = $request->get('ruang');
-        $kelas->matkul_id = $request->get('matkul');
-        $kelas->dosen_id = $request->get('dosen');
+        $user = Auth::user();
+        if($user['status'] == 'admin')
+        {
+            $kelas = Kelas::find($id);
+            $kelas->hari = $request->get('hari');
+            $kelas->hari = $request->get('hari');
+            $kelas->jam_masuk = $request->get('jammasuk');
+            $kelas->jam_keluar = $request->get('jamkeluar');
+            $kelas->kp = $request->get('kp');
+            $kelas->kapasitas = $request->get('kapasitas');
+            $kelas->ruang = $request->get('ruang');
+            $kelas->matkul_id = $request->get('matkul');
+            $kelas->dosen_id = $request->get('dosen');
 
-        $kelas->save();
+            $kelas->save();
 
-        return redirect('master_kelas');
+            return redirect('master_kelas');
+        }
+        else
+        {
+            return view('noaccess');
+        }
+        
     }
 
     /**
@@ -129,10 +175,19 @@ class MasterKelasController extends Controller
      */
     public function destroy($id)
     {
-        $kelas = Kelas::find($id);
+        $user = Auth::user();
+        if($user['status'] == 'admin')
+        {
+            $kelas = Kelas::find($id);
 
-        $kelas->delete();
+            $kelas->delete();
+            
+            return redirect('master_kelas');
+        }
+        else
+        {
+            return view('noaccess');
+        }
         
-        return redirect('master_kelas');
     }
 }
